# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Sumwhere server API",
        "title": "Sumwhere API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://www.sumwhere.kr",
            "email": "qjadn0914@naver.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "www.sumwhere.kr",
    "basePath": "/v1",
    "paths": {
        "/admin/assgin": {
            "patch": {
                "description": "기존 어드민사용자가 추가로 어드민 권한을 유저에게 ≠부여",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "어드민 권한 부여",
                "parameters": [
                    {
                        "description": "{'id':1}",
                        "name": "id",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.AdminController"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "해당 유저정보",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
...
        "/user/all": {
            "get": {
                "description": "가입자 리스트 반환",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "유저 리스트 반환",
                "parameters": [
                    {
                        "description": "순서,순서마다의 정렬순서 (desc|asc), 어디서부터, 몇개를 가져올지",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.GetQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "쿼리의 결과",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/utils.ArrayResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.AdminController": {
            "type": "object"
        },
...
        "models.GetQuery": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 2
                },
                "offset": {
                    "type": "integer",
                    "example": 1
                },
                "orderBy": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "desc"
                    ]
                },
                "sortBy": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "email"
                    ]
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "example@naver.com"
                },
                "gender": {
                    "type": "string"
                },
                "hasProfile": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "joinType": {
                    "type": "string"
                },
                "mainProfileImage": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "example": "1234qwer"
                },
                "point": {
                    "type": "integer"
                },
                "snsId": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "utils.ArrayResult": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "object"
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        }
    }
}